lab10/pyproject.toml
[project]
name = "lab8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-stubs>=6.0.1",
    "djangorestframework>=3.17.1",
]
